Import prebuilt kernel modules like fips140.ko
Add support for copying prebuilt kernel modules like fips140.ko into the
vendor kernel ramdisk. These kernel modules are different in the following
ways:

- They do not appear in vendor_kernel_boot.modules.load provided by the
  kernel build system.
- We want to put their names at the very beginning of the output
  modules.load file to ensure that they are loaded prior to all other
  modules.

We define a new Makefile variable named
BOARD_PREBUILT_VENDOR_KERNEL_RAMDISK_KERNEL_MODULES. For every filename
on this list, we verify that the file exists in $(KERNEL_MODULE_DIR),
and if it does, we include its name in the modules.load for the vendor
kernel boot ramdisk.

Bug: 194712883
Change-Id: I91a1d834b32717d32621d0a67da9052324cab20c